refactor: standalone utils
